Fix for ignoring routes correctly
In Helpers/RouteHelper.php the isExcludedRoute function sets the scope
of the '$isExcluded' variable such that it doesn't get applied correctly
in the array_walk closure. This means that the route never gets excluded
correctly as expected. This can be confirmed by installing
barryvdh/laravel-debugbar and adding 'debugbar' to the 'exclude_route_groups'
config array. If this is done all debugbar routes are never excluded.